Break up resources classes into smaller one
- OwnerResource broken down into OwnerResource, OwnerProductResource, OwnerContentResource
- The tests for it were divided equally
- ConsumerResource broken down into ConsumerResource, GuestIdResource
