+++ src/main/java/org/candlepin/guice/CandlepinModule.java
@@ -119,8 +119,11 @@ import org.candlepin.resource.DeletedConsumerResource;
 import org.candlepin.resource.DistributorVersionResource;
 import org.candlepin.resource.EntitlementResource;
 import org.candlepin.resource.EnvironmentResource;
+import org.candlepin.resource.GuestIdResource;
 import org.candlepin.resource.HypervisorResource;
 import org.candlepin.resource.JobResource;
+import org.candlepin.resource.OwnerContentResource;
+import org.candlepin.resource.OwnerProductResource;
 import org.candlepin.resource.OwnerResource;
 import org.candlepin.resource.PoolResource;
 import org.candlepin.resource.ProductResource;
@@ -323,6 +326,9 @@ public class CandlepinModule extends AbstractModule {
         bind(HypervisorResource.class);
         bind(JobResource.class);
         bind(OwnerResource.class);
+        bind(OwnerProductResource.class);
+        bind(OwnerContentResource.class);
+        bind(GuestIdResource.class);
         bind(PoolResource.class);
         bind(ProductResource.class);
         bind(RoleResource.class);
