+++ src/test/java/org/candlepin/resource/ConsumerResourceCreationTest.java
@@ -224,7 +224,6 @@ public class ConsumerResourceCreationTest {
             this.modelTranslator,
             this.jobManager,
             this.dtoValidator,
-            this.guestIdCurator,
             this.principalProvider,
             this.contentOverrideValidator,
             this.consumerContentOverrideCurator,
