+++ src/test/java/org/candlepin/resource/ConsumerResourceTest.java
@@ -256,7 +256,6 @@ public class ConsumerResourceTest {
             this.translator,
             this.jobManager,
             this.dtoValidator,
-            this.guestIdCurator,
             this.principalProvider,
             this.contentOverrideValidator,
             this.consumerContentOverrideCurator,
@@ -454,7 +453,6 @@ public class ConsumerResourceTest {
             this.translator,
             this.jobManager,
             this.dtoValidator,
-            this.guestIdCurator,
             this.principalProvider,
             this.contentOverrideValidator,
             this.consumerContentOverrideCurator,
