+++ src/test/java/org/candlepin/resource/ConsumerResourceUpdateTest.java
@@ -233,7 +233,6 @@ public class ConsumerResourceUpdateTest {
             this.modelTranslator,
             this.jobManager,
             this.dtoValidator,
-            this.guestIdCurator,
             this.principalProvider,
             this.contentOverrideValidator,
             this.consumerContentOverrideCurator,
